chore: remove cidrpool creation
